requirements.txt
google-cloud-texttospeech==2.2.0
playsound==1.2.2
redis==3.5.3
